Create Result Unit NonUCUM
Physical quantity-type results should always be expressed via a case-sensitive UCUM code. But since EMRs must represent data reported from a lab system which might include discrete units which are NOT specified in UCUM, there should be a clear way to represent a discrete value and unit in CDA that will be interoperable for everyone.